fix(UnicodeTable): character display problem
There's a `Char` (two bytes) array buffer with the length of two held in it; each Unicode codepoint can be represented in UTF-16 as two `Char` type.
Now when dragged too down (codepoint > 0xFFFF), the second char in the buffer will not be zero any more. when drag back to the top, the second non-zero char will mess the display.