fix(rust-build): build message not shown
It's strange. In scala, and at the same time in Gradle (task running phase), in two threads the `stdout` and `stderr` can't be written simultaneously(the Gradle output will display one of them)...

The solution is to put the relevant code in Java, then call them in Scala. Likely it's the problem of Scala `System.out`/`Console.out` and `System.err`/`Console.err` which has different behaviors with Java `System.out` and `System.err`. Weird.