feat(Email): add the basic sending function
JNI: Rust; use the "lettre" crate. The crate needs openssl FFI binding,
so the openssl library should be included in the App "libs". So far the
openssl library must be compiled first, then manually specify the
location of "libcrypto.so" and "libssl.so" in `build.gradle` (works
with `RustBuildPlugin`).

So the build system hasn't been configured completely,
and the CI should be skipped.

[skip ci]