fix(Email): problem that cannot send empty-body emails
Seems the sending progress will be stuck when the subject is empty, and
also sometimes it will be stuck randomly with non-empty subject. It's
weird. I remembered that I had successfully sent an email with empty
subject, and now I can't, because it'll be stuck.

I don't know why it'll be stuck. It it supposed to return an `Err` result
in Rust, and then throw an Java exception, then the blocking will be
released. But now it keeps blocking. In addition I haven't had some way to
debug JNI written with Rust, no knowing what causes the problem.