fix(Wubi): black candidates text in dark mode while the background is also black
Now the candidate view of the Wubi input method will turn to dark mode
(night mode) when the app is in dark mode. But this isn't the original
idea. First, I found the candidates text color won't change
automatically when the editor (the app that is to be committed text from
IME) is in dark mode, keeping in black. I didn't find any solution about
this, and I can't also get the editor's theme (light or dark). So finally
I decided to set background color and text color of the candidate view
following the "Tools" app but not the editor.

Closes #216