TaskNotes: use a hardcoded random notification id
The notification won't be dismissed even if the app
is killed, but the lazy-initialized notification
id will become uninitialized after app is killed,
causing showing multiple notifications.

I don't know why, so a workaround is to set a hard-coded
unique notification id to be used.

Killing methods includes:
- raising SIGKILL
- via `System.exit(0)`

When the user swipe the app activity away in Android
Recents screen, the notification will be gone.
But when that activity in Recents is locked, after
being swiped away, the notification will be still
there. So this may be system-level controlled.