FDB: use `ArrayList` to save `Path`, replacing `LinkedList`
Due to `LinkedList` are through to be inefficient or
even harmful. I read discussions from here:

https://discuss.kotlinlang.org/t/why-kotlin-does-not-provide-linkedlist-implementation/15991.

I tested the operation performing time on `undo()` and `importPathFile()`
the two methods, in both `ArrayList` and `LinkedList` cases, and found
there's no significant difference at the performance. So just follow
these advice and use `ArrayList`.

The reason not using `Stack` is that a `Stack` is based on `Vector`,
and `Vector` is synchronized, this affects the performance.