Revert "TaskNotes: auto show soft keyboard when showing dialog"
This reverts commit ba33e96fb21d83a247c751566b5046cdc466eb3e.

Actually auto showing the soft keyboard
will affect user experience. When there are not only
EditText in the dialog layout, as users, they may
not have an intention to edit the text (maybe interact
with other widgets). So in this case
the soft keyboard shouldn't pop automatically
and there's no need to auto focus the edit field.