TaskNotes: fix wrong item position when viewing all records
When re-ordering items, the order ordinals will be
recorded separately, for each single day. So when
querying all the records and then just order them
by the order ordinal, it will produce wrong results.

Like:

Day 1: 1 2 3 4 5
Day 2: 1 2 3 4 5

Result after the sort: 1 1 2 2 3 3 4 4 5 5, which is
all mixed together.

So the solution now is to order the items
also separately in respect of single days.

But a new issue now is that, it's unable to
re-order
items across different dates, that is, only
can change the
item order in their own dates independently.