FDB: add ".path" default filename in file picker on export
I removed the old logic: when
a name was entered, automatically add the
trailing ".path" extension. But in
737e6f7369dd0ceea75607b0ef37fea3cec3b8f3
file existence check was added: it checks the filename
EditText in FilePicker, without ".path" extension.

For example if I just entered "1" (path file will be
saved as "1.path"), and under the current directory
there is already a "1.path". "1" != "1.path",
duplicated-file-check
won't work. This causes inconsistency.